requirements.txt
aiokafka==0.7.2
kafka-python==2.0.2
librosa==0.9.1
